Refactor and update tests.
